+++ Experiments-iOS-OpenGL/shaders/SimpleFragment.glsl
@@ -1,10 +1,10 @@
-varying lowp vec4 DestinationColor;
-uniform lowp float iTime;
-uniform lowp vec3 iResolution;
+varying highp vec4 DestinationColor;
+uniform highp float iTime;
+uniform highp vec3 iResolution;
 
 //shadertoy function
-void mainImage( out lowp vec4 fragColor, in lowp vec2 fragCoord ) {
-    lowp vec2 uv = fragCoord.xy / iResolution.xy;
+void mainImage( out highp vec4 fragColor, in highp vec2 fragCoord ) {
+    highp vec2 uv = fragCoord.xy / iResolution.xy;
     fragColor = vec4(uv,0.5+0.5*sin(iTime),0.9+0.1*sin(iTime));
 }
